move spec utils to paddlefleet
