+++ src/paddlefleet/transformer/moe/moe_layer.py
@@ -27,7 +27,7 @@ from paddle.distributed.fleet.utils.sequence_parallel_utils import (
 
 if TYPE_CHECKING:
     from paddlefleet.process_groups_config import ProcessGroupCollection
-    from paddlefleet.transformer.spec_utils import LayerSpec
+    from paddlefleet.spec_utils import LayerSpec
     from paddlefleet.transformer.transformer_config import TransformerConfig
 
 from paddlefleet import utils
